Adding license: GPL-3
Adding license file (downloaded from gnu.org) for GPL-3.
Added file 'AUTHORS' with a list of participating authors
that hold the copyright for this software under the GPL-3.
Added a comment on top of the single C source file stating
that this software is licensed under the GNU General Public
License version 3 or any later version.