Out-of-tree builds; refactoring into multiple files
The Makefile will not advise the C compiler to write the
object files into a temporary directory inside /tmp/.
This is meant and beneficial if /tmp is a RAM disk (fast)
and storage location of the C code is a SSD (don't like
many write operations).

The previously single source file was refactored: Several
generic helper functions reside now in utils.c and can be
accessed by including utils.h.